requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
attrs==25.1.0
babel==2.18.0
blinker==1.9.0
cachetools==5.5.2
certifi==2024.12.14
chardet==7.0.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
courlan==1.3.2
cycler==0.12.1
dateparser==1.3.0
distro==1.9.0
filelock==3.17.0
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.7.0
fsspec==2025.3.0
fuzzywuzzy==0.18.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.70.0
h11==0.16.0
h5py==3.13.0
htmldate==1.9.4
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.29.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.10.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jusText==3.0.2
keras==3.8.0
kiwisolver==1.4.8
Levenshtein==0.27.1
libclang==18.1.1
lxml==6.0.2
lxml_html_clean==0.4.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.4.4
namex==0.0.8
narwhals==1.28.0
networkx==3.4.2
newsapi-python==0.2.7
numpy==1.26.4
openai==0.28.1
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.3.2
protobuf==4.25.6
pyarrow==19.0.1
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
PyMuPDF==1.25.3
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.27.1
pytz==2025.1
PyYAML==6.0.2
RapidFuzz==3.12.2
referencing==0.36.2
regex==2024.11.6
reportlab==4.2.5
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==75.8.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.42.2
sympy==1.13.3
tenacity==9.0.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-hub==0.16.1
termcolor==2.5.0
tf_keras==2.16.0
threadpoolctl==3.6.0
tld==0.13.2
tokenizers==0.21.0
toml==0.10.2
torch==2.2.2
tornado==6.4.2
tqdm==4.67.1
trafilatura==1.12.2
transformers==4.49.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xgboost==3.0.0
yarl==1.20.1
